+++ Sources/MapboxSearch/PublicAPI/Use Cases/Discover API/Discover.swift
@@ -63,9 +63,11 @@ extension Discover {
         userActivityReporter.reportActivity(forComponent: "discover-search-nearby")
 
         let searchOptions = SearchOptions(
+            countries: [options.country?.countryCode].compactMap { $0 },
             languages: [options.language.languageCode],
             limit: options.limit,
-            proximity: proximity
+            proximity: proximity,
+            origin: options.origin
         )
 
         search(for: query, with: searchOptions, completion: completion)
@@ -89,10 +91,12 @@ extension Discover {
         userActivityReporter.reportActivity(forComponent: "discover-search-in-area")
 
         let searchOptions = SearchOptions(
+            countries: [options.country?.countryCode].compactMap { $0 },
             languages: [options.language.languageCode],
             limit: options.limit,
-            proximity: proximity,
-            boundingBox: region
+            proximity: proximity ?? options.proximity,
+            boundingBox: region,
+            origin: options.origin
         )
 
         search(for: query, with: searchOptions, completion: completion)
@@ -114,8 +118,11 @@ extension Discover {
         userActivityReporter.reportActivity(forComponent: "discover-search-along-the-route")
 
         let searchOptions = SearchOptions(
+            countries: [options.country?.countryCode].compactMap { $0 },
             languages: [options.language.languageCode],
             limit: options.limit,
+            proximity: options.proximity,
+            origin: options.origin,
             routeOptions: route
         )