+++ Sources/MapboxSearch/PublicAPI/Use Cases/Discover API/Models/Discover+Options.swift
@@ -17,12 +17,31 @@ extension Discover {
         /// English (en) language parameter, but Frankreich (“France”) with a German (de) language parameter.
         public let language: Language
 
+        /// See ``MapboxSearch.Country.ISO3166_1_alpha2`` for the list of ISO 3166 alpha 2 country codes.
+        /// The default value is nil.
+        public let country: Country?
+
+        /// Bias the response to favor results that are closer to a specific location.
+        /// When both proximity and origin are provided, origin is interpreted as the target of a route, while proximity
+        /// indicates the current user location.
+        public let proximity: CLLocationCoordinate2D?
+
+        /// The location from which to calculate distance. When both proximity and origin are provided, origin is
+        /// interpreted as the target of a route, while proximity indicates the current user location.
+        public let origin: CLLocationCoordinate2D?
+
         public init(
             limit: Int = 10,
-            language: Language? = nil
+            language: Language? = nil,
+            country: Country? = nil,
+            proximity: CLLocationCoordinate2D? = nil,
+            origin: CLLocationCoordinate2D? = nil
         ) {
             self.limit = limit
             self.language = language ?? .default
+            self.country = country
+            self.proximity = proximity
+            self.origin = origin
         }
     }
 }