+++ Tests/MapboxSearchUITests/Integration/SearchIntegrationTestCase.swift
@@ -1,6 +1,6 @@
 import XCTest
 
-class SearchIntegrationTestCase: MockServerTestCase {
+class SearchIntegrationTestCase: MockServerUITestCase {
     func testRecentSearchRemove() throws {
         try server.setResponse(.suggestSanFrancisco)
         try server.setResponse(.retrieveSanFrancisco)
@@ -30,6 +30,7 @@ class SearchIntegrationTestCase: MockServerTestCase {
 
     func testSearchCancel() throws {
         try server.setResponse(.suggestSanFrancisco)
+        try server.setResponse(.retrieveSanFrancisco)
 
         app.launch()
         let searchBar = app.searchBar
@@ -40,20 +41,5 @@ class SearchIntegrationTestCase: MockServerTestCase {
         waitForHittable(searchResult.cells["San Francisco"].firstMatch)
         waitForHittable(searchBar.buttons["CancelButton"]).tap()
         XCTAssertFalse(searchResult.exists, "SearchResultTableView Shouldn't exists")
-    }
-
-    func testRecursionSearch() throws {
-        try server.setResponse(.recursion)
-
-        app.launch()
-        let searchBar = app.searchBar
-
-        waitForHittable(searchBar).tap()
-        searchBar.typeText("Recursion")
-
-        let searchResult = waitForHittable(app.mapboxSearchController.searchResultTableView)
-        waitForHittable(searchResult.cells["Did you mean recursion?"].firstMatch).tap()
-        waitForHittable(searchResult.cells["Did you mean recursion?"].firstMatch).tap()
-        waitForHittable(searchResult.cells["Did you mean recursion?"].firstMatch).tap()
     }
 }